ai-kore/pyproject.toml
[project]
name = "ai-kore"
version = "0.1.0"
description = "用于智能表情包生成系统的AI模块"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.129.2",
    "uvicorn>=0.41.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.27.0",
    "oss2>=2.18.0",
    "open-clip-torch>=2.24.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=10.0.0",
    "paddleocr>=2.7.0,<3.0",
    "paddlepaddle>=2.5.0,<3.0",
    "pymilvus>=2.3.0",
    "python-multipart>=0.0.22",
]